@@ -1,13 +1,12 @@
 {
   "name": "mini_program_gizwits_sdk",
-  "version": "1.0.47",
+  "version": "1.0.49",
   "description": "小程序配网sdk",
-  "main": "index.js",
+  "main": "./dist/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "compile": "./node_modules/typescript/bin/tsc",
-    "prepare": "npm run tsc",
-    "tsc": "node ./node_modules/typescript/lib/tsc.js"
+    "test": "node_modules/mocha/bin/_mocha -r ts-node/register tests/**/*.spec.ts -file ./tests/setup-wx.js -timeout 60s",
+    "build": "webpack",
+    "cover": "istanbul cover node_modules/mocha/bin/_mocha -- -r ts-node/register tests/**/*.spec.ts -file ./tests/setup-wx.js -timeout 60s"
   },
   "repository": {
     "type": "git",
@@ -19,7 +18,17 @@
     "sdk"
   ],
   "devDependencies": {
-    "typescript": "^3.3.3333"
+    "@types/chai": "^4.2.11",
+    "@types/mocha": "^7.0.2",
+    "@types/node": "^12.12.14",
+    "chai": "^4.2.0",
+    "istanbul": "^1.1.0-alpha.1",
+    "mocha": "^7.1.1",
+    "ts-loader": "^6.2.2",
+    "ts-node": "^8.8.1",
+    "typescript": "^3.8.3",
+    "webpack": "^4.42.1",
+    "webpack-cli": "^3.3.11"
   },
   "author": "gizwits",
   "license": "ISC",
@@ -28,6 +37,7 @@
   },
   "homepage": "https://github.com/gizwits/mini_program_gizwits_sdk#readme",
   "dependencies": {
-    "@types/node": "^12.12.14"
+    "aes-js": "^3.1.2",
+    "js-md5": "^0.7.3"
   }
 }
